+++ components/components.ts
@@ -48,3 +48,4 @@ export * from './upload';
 export * from './virtual-list';
 export * from './text';
 export * from './timeline';
+export * from './calendar';