+++ components/select-cascader/const.ts
@@ -1,0 +1,3 @@
+export const SELECT_CASCADER_NAME = 'FSelectCascader' as const;
+
+export const LABEL_SEPARATOR = ' / ';