+++ components/select/const.ts
@@ -1,4 +1,8 @@
+import getPrefixCls from '../_util/getPrefixCls';
+
 export const SELECT_PROVIDE_KEY = Symbol('FSelect');
 
 export const PADDING_LEFT_BASE = 8;
 export const PADDING_LEFT_INDENT = 16;
+
+export const prefixCls = getPrefixCls('select');