+++ docs/.vitepress/components/calendar/modeMonth.vue
@@ -1,0 +1,3 @@
+<template>
+    <FCalendar mode="month" />
+</template>