+++ docs/.vitepress/components/select/multiple.vue
@@ -2,14 +2,14 @@
     <FForm :labelWidth="200">
         <FFormItem label="选中项展示是否有边框：">
             <FRadioGroup v-model="tagBordered">
-                <FRadio :value="false">false(默认)</FRadio>
-                <FRadio :value="true">true</FRadio>
+                <FRadio :value="false">否(默认)</FRadio>
+                <FRadio :value="true">是</FRadio>
             </FRadioGroup>
         </FFormItem>
         <FFormItem label="多选是否折叠展示：">
             <FRadioGroup v-model="collapseTags">
-                <FRadio :value="false">false(默认)</FRadio>
-                <FRadio :value="true">true</FRadio>
+                <FRadio :value="false">否(默认)</FRadio>
+                <FRadio :value="true">是</FRadio>
             </FRadioGroup>
         </FFormItem>
         <FFormItem v-if="collapseTags" label="折叠项限制：">