+++ docs/.vitepress/configs/index.ts
@@ -1,0 +1,4 @@
+import navbar from './navbar/index';
+import sidebar from './sidebar/index';
+
+export { navbar, sidebar };