+++ scripts/genIcons.mjs
@@ -1,30 +1,18 @@
-/* eslint import/no-extraneous-dependencies: 0 */
-
 import { join, basename } from 'path';
 import { readdirSync, writeFileSync } from 'fs';
 
 import optimizeSvg from './optimizeSvg.mjs';
-import { stringToCamelCase } from './utils.mjs';
+import { getProjectRootDir, stringToCamelCase } from './utils.mjs';
+import { SVG_COMPONENT_TMPLATE } from './constants.mjs';
 
-const outputPath = join(process.cwd(), './components/icon');
+const rootDir = getProjectRootDir();
+const outputPath = join(rootDir, './components/icon');
 const exportAllIconPath = join(
-    process.cwd(),
+    rootDir,
     './docs/.vitepress/theme/IconDoc/icons.js',
 );
-const iconFileBase = join(process.cwd(), './icons');
+const iconFileBase = join(rootDir, './icons');
 const iconFiles = readdirSync(iconFileBase);
-
-const SVG_COMPONENT_TMPLATE = `
-import IconWrapper from './IconWrapper';
-import type { IconProps } from './IconWrapper';
-import './style';
-
-export default (props?: IconProps) => (
-    <IconWrapper {...props} ATTRS>
-        SVG
-    </IconWrapper>
-);
-`;
 
 function genExportAllIconFile(iconNames) {
     const content = iconNames
@@ -49,7 +37,6 @@ function gen() {
         iconFiles.map((item) => join(iconFileBase, item)),
     );
     const iconNames = [];
-    // eslint-disable-next-line
     for (const { fileName, data } of svgDatas) {
         const iconName = stringToCamelCase(basename(fileName, '.svg'));
         iconNames.push(iconName);